Make file tree open files in first column
