+++ app/gist/template.hbs
@@ -34,12 +34,9 @@
   {{#unless fullScreen}}
     {{#if fileTreeShown}}
       <div class="col-md-4 file-tree">
-        <div class="header">
-          <div class="nav-right">
-            <span class="glyphicon glyphicon-chevron-left" {{action 'hideFileTree'}}></span>
-          </div>
-        </div>
-        {{ember-jstree data=fileTreeData}}
+        {{file-tree model=model
+                    openFile=(action "openFile")
+        }}
       </div>
     {{/if}}
