+++ app/styles/app.scss
@@ -33,6 +33,14 @@ body {
   position: absolute;
 }
 
+.light-gray {
+  color: #cccccc;
+}
+
+.yellow {
+  color: #ffcc00;
+}
+
 .toolbar {
   height: $topbar-height;
   font-size: 13px;
