+++ app/templates/components/file-tree.hbs
@@ -1,0 +1,9 @@
+<div class="header">
+  <div class="nav-right">
+    <span class="glyphicon glyphicon-chevron-left" {{action 'hideFileTree'}}></span>
+  </div>
+</div>
+{{ember-jstree data=fileTreeData
+               actionReceiver=jsTreeActionReceiver
+               eventDidSelectNode="handleSelectTreeNode"
+}}
